% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_taxon_names.R
\name{clean_taxon_names}
\alias{clean_taxon_names}
\title{clean_taxon_names()}
\usage{
clean_taxon_names(data, target_taxon, life_stage, database = "gbif")
}
\arguments{
\item{data}{\itemize{
\item data.frame with a column containing target taxon names and
life-stages
}}

\item{target_taxon}{\itemize{
\item character string with the column name containing the
taxon names
}}

\item{life_stage}{\itemize{
\item character string with the column name containing the
life_stages
}}

\item{database}{\itemize{
\item taxonomic database to use: gbif (default), itis, col
}}
}
\value{
data.frame with target taxon names and cleaned names for the chosen
taxonomic backbone
}
\description{
Get taxonomic distances of target names relative to the taxa
databases
}
\details{
This function takes a data.frame with target_taxa and uses the bdc
package to clean and harmonise the names.
}
\author{
James G. Hagan (james_hagan(at)outlook.com) and Ronald Bergmann
}
