% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_database}
\alias{download_database}
\title{download_database}
\usage{
download_database(pkgname)
}
\description{
Will setup the local database in a directory
chosen by rappdirs. Will download DB files for the current
package version (tag) if it can be determined and is available
or from mainline otherwise.
}
