% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_traits_tax_dist.R
\name{extract_body_size_range_match}
\alias{extract_body_size_range_match}
\title{extract_body_size_range_match()}
\usage{
extract_body_size_range_match(
  equation_id,
  target_body_size,
  prop = 0.3,
  equation_db
)
}
\arguments{
\item{equation_id}{\itemize{
\item id number of the equation
}}

\item{target_body_size}{\itemize{
\item body size of the target taxon
}}

\item{prop}{\itemize{
\item proportion above and below equation body size range
considered acceptable for using an equation
}}

\item{equation_db}{\itemize{
\item equation database object
}}
}
\value{
string with the genus name
}
\description{
calculate whether target length matches equation range
}
\details{
This function generates a data.frame with three variables:
body_size_range_match: whether the target length is within the equation
length range (TRUE or FALSE), body_size_min_dist: percentage away
from the minimum length used to create the equation, body_size_max_dist:
percentage away from the maximum length.
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
