% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_traits_tax_dist.R
\name{extract_genus}
\alias{extract_genus}
\title{extract_genus()}
\usage{
extract_genus(binomial)
}
\arguments{
\item{binomial}{\itemize{
\item binomial character string separated by a space (e.g.
"Loxodonta loxodonta")
}}
}
\value{
string with the genus name
}
\description{
If a binomial taxa is supplied, extract the genus name only
}
\details{
This method works exclusively with genera and not species.
This is because once the genus is known, the relationship among the
species within that genus is known. It makes the method much more
computationally efficient. This function is used to then extract the genus
from a species name. The genus name is used to place the species in the
correct taxonomic framework
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
