% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_habitat_data.R
\name{get_habitat_data}
\alias{get_habitat_data}
\title{get_habitat_data()}
\usage{
get_habitat_data(data, latitude_dd, longitude_dd)
}
\arguments{
\item{data}{\itemize{
\item data.frame with a column containing latitude and longitude in
decimal degrees
}}

\item{latitude_dd}{\itemize{
\item character string with the column name containing the
latitude in decimal degrees
}}

\item{longitude_dd}{\itemize{
\item character string with the column name containing the
longitude in decimal degrees
}}
}
\value{
tibble of the input data with habitat data from Abell et al.s (2008)
ecoregion map attached as additional columns function to get habitat data
}
\description{
Get habitat data from Abell et al.'s (2008) freshwater
ecoregion map
}
\details{
This function takes a data.frame with latitude and longitude data
and gets data on biogeographic realm, major habitat type and ecoregion
based on Abell et al.'s (2008) freshwater ecoregion map.
}
\author{
James G. Hagan (james_hagan(at)outlook.com) and Ronald Bergmann
}
