% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_from_taxon.R
\name{get_trait_from_taxon}
\alias{get_trait_from_taxon}
\title{get_trait_from_taxon()}
\usage{
get_trait_from_taxon(
  data,
  target_taxon,
  life_stage,
  latitude_dd,
  longitude_dd,
  body_size,
  workflow = "workflow2",
  max_tax_dist = 3,
  trait = "equation",
  gen_sp_dist = 0.5
)
}
\arguments{
\item{data}{\itemize{
\item data.frame with at least five columns: target taxon, life
stage, latitude (dd), longitude (dd) and body size (mm)
if trait == "equation"
}}

\item{target_taxon}{\itemize{
\item character string with the column name containing the
taxon names
}}

\item{life_stage}{\itemize{
\item character string with the column name containing the
life-stage information
}}

\item{latitude_dd}{\itemize{
\item character string with the column name containing the
latitude in decimal degrees
}}

\item{longitude_dd}{\itemize{
\item character string with the column name containing the
longitude in decimal degrees
}}

\item{body_size}{\itemize{
\item character string with the column name containing the
body size data if trait = "equation"
}}

\item{workflow}{\itemize{
\item options are "workflow1" or "workflow2"
(default = "workflow2)
}}

\item{max_tax_dist}{\itemize{
\item maximum taxonomic distance acceptable between the
target and the taxa in the database (default = 3)
}}

\item{trait}{\itemize{
\item trait to be searched for (default = "equation")
}}

\item{gen_sp_dist}{\itemize{
\item taxonomic distance between a genus and a species
(default = 0.5)
}}
}
\value{
tibble with chosen traits or equations based on the input parameters
}
\description{
Use taxonomic distance, life-stages, habitat data etc. to
select an appropriate trait/equation
}
\details{
This function uses the \code{\link[=clean_taxon_names]{clean_taxon_names()}} function to clean and
harmonise the names of a target dataset. The \code{\link[=get_habitat_data]{get_habitat_data()}} function is
then used to get relevant habitat information. The
\code{\link[=select_traits_tax_dist]{select_traits_tax_dist()}} function calculates taxonomic distance for each
focal taxon and the relevant entries in the different higher-level taxon
graphs generated from the GBIF, ITIS and COL taxonomic backbones. This
information is combined to choose an appropriate trait or equation for each
target taxon name.
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
