% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_traits_tax_dist.R
\name{select_traits_tax_dist}
\alias{select_traits_tax_dist}
\title{select_traits_tax_dist()}
\usage{
select_traits_tax_dist(
  data,
  target_taxon,
  body_size,
  life_stage,
  max_tax_dist = 3,
  trait = "equation",
  gen_sp_dist = 0.5
)
}
\arguments{
\item{data}{\itemize{
\item input data.frame exported from \code{\link[=get_habitat_data]{get_habitat_data()}} and
\code{\link[=clean_taxon_names]{clean_taxon_names()}} function
}}

\item{target_taxon}{\itemize{
\item character string with the column name containing the
taxon names
}}

\item{body_size}{\itemize{
\item column name containing the body length data
}}

\item{life_stage}{\itemize{
\item character string with the column name containing the
life-stage information
}}

\item{max_tax_dist}{\itemize{
\item maximum taxonomic distance acceptable between the
target and the taxa in the database (default = 3)
}}

\item{trait}{\itemize{
\item trait to be searched for (default = "equation")
}}

\item{gen_sp_dist}{\itemize{
\item taxonomic distance between a genus and a species
(default = 0.5)
}}
}
\value{
tibble of the input data with traits or equations within the maximum
taxonomic distance
}
\description{
Get taxonomic distances of target names relative to the taxa
databases
}
\details{
This function searches the relevant trait or equation database for
the best matching trait or equation for a given target name based on
taxonomic distance
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
