% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_names.R
\name{special_taxon_names}
\alias{special_taxon_names}
\title{Special taxon names}
\usage{
special_taxon_names()
}
\value{
vector of special taxon names
}
\description{
A vector of special taxon names that are not searched using
the regular search framework
}
\details{
Some taxa are often not properly identified in freshwater field
research. For example, different worm taxa, tardigrads, rotifers etc. are
often specified as simply: Oligochaeta, Rotifera, Nematoda etc. Thus,
these special names can not be used in the more general taxonomic
similarity framework and are rather matched to a specific set of equations
developed for those taxa. We provide several of these for different
geographical locations in the database so that they can be matched as
accurately as possible.
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
