% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_user_db}
\alias{update_user_db}
\title{update_user_db}
\usage{
update_user_db()
}
\description{
Copy over DB files to appdata dir. Comes in
handy in case they got stale and you want to update them.
}
